Give run identity its own file
SagaObserver.kt declared five things and was named after the last of
them. Two did not belong there at all.

SagaOutcome is internal control flow with no observer involvement -
it exists only as failed's return channel - and its sibling
StepOutcome already lived in Saga.kt as a file-private type. Two
carriers of the same kind, in different files, at different
visibilities. SagaOutcome now sits beside StepOutcome and is private
too, which is the visibility it should always have had: being in the
wrong file is what forced it wider.

SagaRun and SagaRunId are the machine's per-attempt context, not
observer types. Saga.kt refers to SagaRun fifteen times against the
observer API's seven, threading it through advance, resolveFirst,
resolveNext, failed, terminated and recordRestart as the bundle of
principal, params, identity and attempt. It would exist with no
observers at all; notifications merely expose it. Someone asking what
a SagaRunId is had to know to look in the observer file - and it is
the type the saga_events rows will key on, referenced from hc-server
with no observer in view.

SagaObserver.kt is left with SagaNotification and SagaObserver, which
are one concept and match the name.
